Fix release review issues and trim initial bundle
