+++ src/locales/pt/translation.json
@@ -539,24 +539,24 @@
       "role_producer": "Produtor",
       "original_artist": "Artista original",
       "collaborator": "Colaborador",
-      "cover_of": "Versao de",
+      "cover_of": "Versão de",
       "by": "por",
       "type_label": "Tipo",
       "date_label": "Lançado em",
       "duration_label": "Duração",
-      "country_label": "Pais",
+      "country_label": "País",
       "status_label": "Status",
-      "barcode_label": "Codigo de barras",
-      "musicbrainz_id_label": "ID do lancamento no MusicBrainz",
-      "facts_title": "Ficha do lancamento",
+      "barcode_label": "Código de barras",
+      "musicbrainz_id_label": "ID do lançamento no MusicBrainz",
+      "facts_title": "Ficha do lançamento",
       "tracks": "Faixas",
       "embed_title": "Ouvir {{name}}",
       "musicbrainz_ready_title": "Notas para MusicBrainz",
-      "recording_title_label": "Titulo da gravacao",
-      "artist_credit_label": "Credito de artista",
+      "recording_title_label": "Título da gravação",
+      "artist_credit_label": "Crédito de artista",
       "year_label": "Ano",
-      "genre_label": "Genero",
-      "musicbrainz_ready_note": "Use estes campos como ponto de partida ao conferir ou melhorar o lancamento no MusicBrainz. Mantenha apenas dados verificados em plataformas ou distribuidoras.",
+      "genre_label": "Gênero",
+      "musicbrainz_ready_note": "Use estes campos como ponto de partida ao conferir ou melhorar o lançamento no MusicBrainz. Mantenha apenas dados verificados em plataformas ou distribuidoras.",
       "not_found": "Lançamento não encontrado.",
       "not_found_back": "Voltar ao Hub de Música"
     },
