+++ src/pages/NewsPage.tsx
@@ -18,7 +18,8 @@ import { ARTIST } from '../data/artistData';
 import { sanitizeHtml, safeUrl } from '../utils/sanitize';
 import { stripHtml } from '../utils/text';
 import { getDateTimeFormatter } from '../utils/date';
-import { findReleaseByNewsSlug, getReleaseOpenGraphAlt, getReleaseOpenGraphType } from '../utils/openGraph';
+import { getReleaseOpenGraphAlt } from '../utils/openGraph';
+import { findReleaseByNewsSlug, getReleaseOpenGraphType } from '../utils/releaseOpenGraph';
 import { buildYouTubeVideoObject } from '../utils/youtube';
 import NotFoundPage from './NotFoundPage';
 import { logger } from '../lib/logger';
